requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.2.0
pix2tex[gui]==0.1.2
torch==2.1.0
torchvision==0.16.0
